同步最新代码

+++ src/knockout/koquestion_file.ts
@@ -1,27 +1,34 @@
 import * as ko from "knockout";
-import { JsonObject } from "../jsonobject";
+import { Serializer } from "../jsonobject";
 import { QuestionFactory } from "../questionfactory";
 import { QuestionFileModel } from "../question_file";
 import { QuestionImplementor } from "./koquestion";
 import { Question } from "../question";
+import { confirmAction } from "../utils/utils";
 
 export class QuestionFileImplementor extends QuestionImplementor {
-  koState = ko.observable<string>("empty");
-  koHasValue = ko.computed(() => this.koState() === "loaded");
-  koData = ko.computed(() => {
+  koState: any = ko.observable<string>("empty");
+  koHasValue: any = ko.computed(() => this.koState() === "loaded");
+  koData: any = ko.computed(() => {
     if (this.koHasValue()) {
       return (<QuestionFileModel>this.question).previewValue;
     }
     return [];
   });
-  koInputTitle = ko.observable<string>();
-
+  koInputTitle: any = ko.observable<string>();
+  koChooseFileClass: any = ko.pureComputed(() => {
+    return (
+      this.question.koCss().chooseFile +
+      (this.question.isReadOnly ? " " + this.question.koCss().disabled : "")
+    );
+  });
   constructor(question: Question) {
     super(question);
     var self = this;
     (<any>this.question)["koData"] = this.koData;
     (<any>this.question)["koHasValue"] = this.koHasValue;
     (<any>this.question)["koInputTitle"] = this.koInputTitle;
+    (<any>this.question)["koChooseFileClass"] = this.koChooseFileClass;
     var updateState = (state: any) => {
       this.koState(state);
       this.koInputTitle((<QuestionFileModel>this.question).inputTitle);
@@ -29,17 +36,35 @@ export class QuestionFileImplementor extends QuestionImplementor {
     (<QuestionFileModel>this.question).onStateChanged.add((sender, options) => {
       updateState(options.state);
     });
+    (<any>this.question)["ondrop"] = (data: any, event: any) => {
+      event.preventDefault();
+      let src = event.originalEvent
+        ? event.originalEvent.dataTransfer
+        : event.dataTransfer;
+      this.onChange(src);
+    };
+    (<any>this.question)["ondragover"] = (data: any, event: any) => {
+      event.preventDefault();
+    };
     (<any>this.question)["dochange"] = (data: any, event: any) => {
       var src = event.target || event.srcElement;
       self.onChange(src);
     };
     (<any>this.question)["doclean"] = (data: any, event: any) => {
       var src = event.target || event.srcElement;
+      if (question.needConfirmRemoveFile) {
+        var isConfirmed = confirmAction(question.confirmRemoveAllMessage);
+        if (!isConfirmed) return;
+      }
       var input = src.parentElement.querySelectorAll("input")[0];
       (<QuestionFileModel>this.question).clear();
       input.value = "";
     };
     (<any>this.question)["doremovefile"] = (data: any, event: any) => {
+      if (question.needConfirmRemoveFile) {
+        var isConfirmed = confirmAction(question.getConfirmRemoveMessage(data.name));
+        if (!isConfirmed) return;
+      }
       (<QuestionFileModel>this.question).removeFile(data);
     };
   }
@@ -47,7 +72,10 @@ export class QuestionFileImplementor extends QuestionImplementor {
     if (!(<any>window)["FileReader"]) return;
     if (!src || !src.files || src.files.length < 1) return;
     let files = [];
-    for (let i = 0; i < src.files.length; i++) {
+    let allowCount = (<QuestionFileModel>this.question).allowMultiple
+      ? src.files.length
+      : 1;
+    for (let i = 0; i < allowCount; i++) {
       files.push(src.files[i]);
     }
     src.value = "";
@@ -62,7 +90,7 @@ export class QuestionFile extends QuestionFileModel {
   }
 }
 
-JsonObject.metaData.overrideClassCreatore("file", function() {
+Serializer.overrideClassCreator("file", function() {
   return new QuestionFile("");
 });
 QuestionFactory.Instance.registerQuestion("file", name => {
